Fixed starwars schemas
