+++ examples/starwars-ts/schema.ts
@@ -1,101 +1,104 @@
 import {
   ObjectType,
+  InterfaceType,
   Schema,
   Field,
-  Interface,
-  String,
   ID,
-  List,
-  Enum
-} from '../../src/types_old';
+  description,
+  deprecated,
+  EnumType,
+  NonNull
+  // EnumValue
+  // List,
+  // Enum
+} from '../../src/types';
 import { getHero, getFriends, getHuman, getDroid } from './data';
 
-export class Episode extends Enum {
-  static description = 'The description';
-  static values = {
-    NEWHOPE: {
-      value: 4,
-      description: 'Released in 1977.'
-    },
-    EMPIRE: {
-      value: 5,
-      description: 'Released in 1980.'
-    },
-    JEDI: {
-      value: 6,
-      description: 'Released in 1983.'
-    }
-  };
+@EnumType()
+@description('The description')
+export class Episode {
+  @description('Released in 1977.') static NEWHOPE = 4;
+  @description('Released in 1980.') static EMPIRE = 5;
+  @description('Released in 1983.') static JEDI = 6;
 }
 
-class Character extends Interface {
-  static description = 'A character in the Star Wars Trilogy';
-  static fields = {
-    id: new ID({ required: true, description: 'The id of the character.' }),
-    name: new String({ description: 'The name of the character.' }),
-    friends: new List(Character, {
-      description:
-        'The friends of the character, or an empty list if they have none.'
-    }),
-    appearsIn: new Episode({ description: 'Which movies they appear in.' })
-  };
-  static resolveType = (root: TCharacter) => {
+@InterfaceType({
+  resolveType: (root: Character) => {
     return getHuman(root.id) ? Human : Droid;
-  };
-
-  friends(root: TCharacter): TCharacter[] {
-    return getFriends(root);
   }
+})
+@description('A character in the Star Wars Trilogy')
+export class Character {
+  @Field(NonNull(ID))
+  @description('The id of the character.')
+  @deprecated('This is no longer used')
+  public id: string;
+
+  @Field(String)
+  @description('The name of the character.')
+  public name: string;
+
+  @Field([Character])
+  @description(
+    'The friends of the character, or an empty list if they have none.'
+  )
+  friends?(): Character[] {
+    return getFriends(this);
+  }
+
+  @Field([Episode])
+  @description('Which movies they appear in.')
+  public appearsIn: number[];
+
+  public friendIds: string[];
 }
 
-type TCharacter = Human | Droid;
-
-export class Human extends ObjectType {
-  static description = 'A humanoid creature in the Star Wars universe.';
-  static interfaces = [Character];
-  static fields = {
-    homePlanet: new String({
-      description: 'The home planet of the human, or null if unknown..'
-    })
-  };
+@ObjectType({
+  interfaces: [Character]
+})
+@description('A humanoid creature in the Star Wars universe.')
+export class Human implements Character {
+  @Field(String)
+  @description('The home planet of the human, or null if unknown..')
+  public homePlanet?: string;
 
   id: string;
   name: string;
+  friends?: () => Character[];
   friendIds: string[];
   appearsIn: number[];
-  homePlanet?: string;
 }
 
-export class Droid extends ObjectType {
-  static description = 'A mechanical creature in the Star Wars universe.';
-  static interfaces = [Character];
-  static fields = {
-    primaryFunction: new String({
-      description: 'The primary function of the droid, or null if unknown..'
-    })
-  };
+@ObjectType({
+  interfaces: [Character]
+})
+@description('A mechanical creature in the Star Wars universe.')
+export class Droid implements Character {
+  @Field(String)
+  @description('The primary function of the droid, or null if unknown..')
+  public primaryFunction: string;
 
   id: string;
   name: string;
+  friends?: () => Character[];
   friendIds: string[];
   appearsIn: number[];
-  primaryFunction?: string;
 }
 
-class Query extends ObjectType {
-  static fields = {
-    hero: new Field(Character, { args: { episode: new Episode() } }),
-    human: new Field(Human, { args: { id: new ID() } }),
-    droid: new Field(Droid, { args: { id: new ID() } })
-  };
-
-  hero(root: any, args: { episode: any }): TCharacter {
+@ObjectType()
+class Query {
+  @Field(Character, { args: { episode: Episode } })
+  hero(args: { episode: number }): Character {
     return getHero(args.episode);
   }
-  human(root: any, args: { id: string }): Human {
+
+  @Field(Human, { args: { id: String } })
+  human(args: { id: string }): Human {
     return getHuman(args.id);
   }
-  droid(root: any, args: { id: string }): Droid {
+
+  @Field(Droid, { args: { id: String } })
+  droid(args: { id: string }): Droid {
     return getDroid(args.id);
   }
 }
@@ -103,5 +106,14 @@ class Query extends ObjectType {
 const schema = new Schema({
   query: Query
 });
+
+// schema
+//   .execute(
+//     `{ hero {
+//   id
+// }
+// }`
+//   )
+//   .then(resp => console.log(resp));
 
 export default schema;
